+++ safe-app/src/test/java/net/maidsafe/api/DefaultContainers.java
@@ -1,0 +1,15 @@
+package net.maidsafe.api;
+
+public final class DefaultContainers {
+    public static final String DOCUMENTS = "_documents";
+    public static final String DOWNLOADS = "_downloads";
+    public static final String MUSIC = "_music";
+    public static final String PICTURES = "_pictures";
+    public static final String VIDEOS = "_videos";
+    public static final String PUBLIC = "_public";
+    public static final String PUBLICNAMES = "_publicNames";
+
+    private DefaultContainers() {
+        // Constructor intentionally empty
+    }
+}